Final test version

+++ include/inets/carrier_sensing_cpp_cc.h
@@ -47,7 +47,7 @@ namespace gr {
        * class. inets::carrier_sensing_cpp_cc::make is the public interface for
        * creating new instances.
        */
-      static sptr make(int develop_mode, int block_id, float cs_duration, float cs_threshold, int system_time_granularity_us);
+      static sptr make(int develop_mode, int block_id, double cs_duration, float cs_threshold, int system_time_granularity_us);
     };
 
   } // namespace inets
